vdev_id: new slot type SES.
This extends vdev_id to support a new slot type, SES, for SCSI Enclosure
Services.  With slot type SES, the disk slot numbers are determined by
using the device slot number reported by sg_ses for the device with
matching SAS address, found by querying all available enclosures.

This is primarily of use on systems with a deficient driver omitting
support for bay_identifier in /sys/devices.  In my testing, I found that
the existing slot types of port and id were not stable across disk
replacement, so an alternative was required.

Signed-off-by: Simon Guest <simon.guest@tesujimath.org>